text_analyser/requirements.txt
spacy
google-cloud-storage
